integration test stability improvements
